[Bronze V] Title: 꼬마 정민, Time: 12 ms, Memory: 856 KB -BaekjoonHub

+++ 백준/Bronze/11382. 꼬마 정민/꼬마 정민.go
@@ -3,7 +3,7 @@ package main
 import "fmt"
 
 func main() {
-	var a, b, c uint64
+	var a, b, c int
 	fmt.Scan(&a, &b, &c)
 	fmt.Println(a + b + c)
 }
